+++ library/Fission/User/Creator/Error.hs
@@ -1,10 +1,9 @@
--- | TODO make me pluckable
 module Fission.User.Creator.Error (Error (..)) where
 
 import           Servant.Server
 
 import           Fission.Prelude
-import qualified Fission.Web.Error     as Web -- TODO move to .Class and drop qualified
+import           Fission.Web.Error
 import qualified Fission.Internal.UTF8 as UTF8
 
 data Error
@@ -16,9 +15,9 @@ data Error
            )
 
 instance Display Error where
-  display FailedDigest = "Could not create password digest"
+  display FailedDigest  = "Could not create password digest"
   display AlreadyExists = "The username or email already exists in our system"
 
-instance Web.ToServerError Error where
+instance ToServerError Error where
   toServerError FailedDigest  = err500 { errBody = UTF8.showLazyBS FailedDigest }
   toServerError AlreadyExists = err409 { errBody = UTF8.showLazyBS AlreadyExists }
